pyproject.toml
[tool.poetry]
name = "datavis-cli"
version = "0.1.1"
description = "Output data loading and visualization boilerplate"
authors = ["Ben Skubi <bskubi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pandas = "^2.2.0"
pyyaml = "^6.0.1"
seaborn = "^0.13.2"

[tool.poetry.scripts]
datavis-cli = "data_vis_cli.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
